Split PieceMoveNotation from PieceMoveGenerator
