+++ tests/wac.rs
@@ -9,6 +9,7 @@ use littlewing::clock::Clock;
 use littlewing::fen::FEN;
 use littlewing::game::Game;
 use littlewing::piece_move_generator::PieceMoveGenerator;
+use littlewing::piece_move_notation::PieceMoveNotation;
 use littlewing::search::Search;
 
 #[test]
